Put convenience functions in overlay/index.js
This moves the interpretation of overlays into overlay/compile.js, and
provides convenience functions for running them in overlay/index.js.

The convenience functions really just specialise `compile()` with the
std library; there is no attempt to generalise the expected format of
the overlay files/objects.
